fix test utils
