+++ python/ray/tests/py3_test.py
@@ -8,7 +8,7 @@ import pytest
 
 import ray
 import ray.cluster_utils
-import ray.tests.utils
+import ray.test_utils
 
 
 @pytest.mark.parametrize(
