+++ python/ray/tests/test_autoscaler_yaml.py
@@ -7,7 +7,7 @@ import unittest
 import yaml
 
 from ray.autoscaler.autoscaler import fillout_defaults, validate_config
-from ray.tests.utils import recursive_fnmatch
+from ray.test_utils import recursive_fnmatch
 
 RAY_PATH = os.path.abspath(os.path.join(__file__, "../../"))
 CONFIG_PATHS = recursive_fnmatch(
