+++ python/ray/tests/test_component_failures.py
@@ -14,8 +14,8 @@ import pytest
 import ray
 import ray.ray_constants as ray_constants
 from ray.cluster_utils import Cluster
-from ray.tests.utils import (run_string_as_driver_nonblocking,
-                             RayTestTimeoutException)
+from ray.test_utils import (run_string_as_driver_nonblocking,
+                            RayTestTimeoutException)
 
 
 # This test checks that when a worker dies in the middle of a get, the plasma
