+++ python/ray/tests/test_dynres.py
@@ -7,7 +7,7 @@ import time
 
 import ray
 import ray.cluster_utils
-import ray.tests.utils
+import ray.test_utils
 
 logger = logging.getLogger(__name__)
