+++ python/ray/tests/test_node_manager.py
@@ -3,7 +3,7 @@ from __future__ import division
 from __future__ import print_function
 
 import ray
-from ray.tests.utils import run_string_as_driver
+from ray.test_utils import run_string_as_driver
 
 
 # This tests the queue transitions for infeasible tasks. This has been an issue
